move to remote

+++ src/evaluate/kfold_EALSTM_tune.py
@@ -121,8 +121,8 @@ tst_rmse_per_ep = np.empty((n_folds,int(n_eps/10)+1))
 # tst_rmse_per_ep = []
 n_hid_arr = np.array([32,64,128,256])
 best_ep_per_hid = np.empty_like(n_hid_arr)
-best_tstrmse_per_hid = np.empty_like(n_hid_arr)
-best_trnrmse_per_hid = np.empty_like(n_hid_arr)
+best_tstrmse_per_hid = np.empty_like(n_hid_arr,dtype=np.float)
+best_trnrmse_per_hid = np.empty_like(n_hid_arr,dtype=np.float)
 best_ep_per_hid[:] = np.nan
 best_tstrmse_per_hid[:] = np.nan
 best_trnrmse_per_hid[:] = np.nan
